Advance_Sql_Agent/requirements.txt
sqlalchemy>=1.4
pyodbc
python-dotenv
google-cloud-aiplatform

fastapi
uvicorn[standard]
jinja2
plotly
pandas